test(shell): consolidate React and Angular Playwright E2E specs into unified parameterized spec
- Delete ng-basic-catalog-integration.e2e.ts and react-basic-catalog-integration.e2e.ts to remove copy-paste redundancy.
- Implement basic-catalog-integration.e2e.ts that accepts ports, renderers, date formatting, and element fill actions mapping to loop dynamically over React and Angular implementations.

TAG=agy
CONV=314fa802-b607-4435-b4e8-ce7319b41f87
